backend/requirements.txt
asgiref==3.2.3
attrs==19.3.0
Django==3.0
django-cors-headers==3.2.0
djangorestframework==3.11.0
more-itertools==8.0.2
packaging==19.2
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.0
pyparsing==2.4.5
pytest==5.3.2
python-decouple==3.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
wcwidth==0.1.7
